requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
gspread
gspread-dataframe
google-api-python-client
google-auth